Almost finished

+++ app.py
@@ -32,10 +32,9 @@ def create_item(text, priority):
 def read_items():
     return db.session.query(Item).order_by(Item.priority).all()
     
-def update_item(item_id, text, priority, done):
+def update_item(item_id, text, done):
     db.session.query(Item).filter_by(id = item_id).update({
 		"text": text,
-        "priority": priority,
 		"done": True if done == "on" else False #checkbox
 	})
     db.session.commit()
@@ -65,19 +64,22 @@ def about():
 @app.route("/edit/<item_id>", methods = ["POST", "GET"])
 def edit_item(item_id):
     if request.method == "POST":
-        update_item(item_id, text=request.form['text'], done=request.form['done'], priority=request.form.get('priority'))
+        update_item(item_id, text=request.form['text'], done=request.form['done'])
     elif request.method == "GET":
         delete_item(item_id)
     return redirect("/", code=302)
 
-@app.route("/delete/<item_id>")
+@app.route("/delete/<item_id>", methods = ["POST", "GET"])
 def delete(item_id):
-    delete_item(item_id)
-    return redirect("/")
+    if request.method == "POST":
+        delete_item(item_id)
+    return redirect("/", code=303)
 
+@app.route("/priority/<item_id>",methods=["POST","GET"])
 def edit_priority(item_id):
     if request.method=="POST":
-        update_item_priority(item_id,request.form['priority'])
+        update_item_priority(item_id,request.form.get('priority'))
+    return redirect("/", code=304)
 
 if __name__ == "__main__":
     db.create_all()
